refactor: Rename 'interstellar' package to 'stellarstructure'
This commit renames the core `interstellar` package to `stellarstructure`. All related classes and imports have been updated to reflect this change. This includes `InterstellarBot` becoming `StellarBot`, and the `interstellar.directives` package being moved to `stellarstructure.runnables`.

The `Directive` class has been split into a new base `Runnable` class and a `Directive` class which extends it. This change introduces the concept of a `Procedure`, a new type of `Runnable` that can execute a sequence of `Directive`s.

Additionally, the Gradle wrapper has been upgraded from version 8.11.1 to 8.13, and the Android Gradle Plugin has been updated from 8.9.1 to 8.13.0.